+++ reikna/cluda/__init__.py
@@ -1,6 +1,6 @@
 from reikna.cluda.api_discovery import *
 from reikna.cluda.api_tools import *
-from reikna.cluda.kernel import Module
+from reikna.cluda.kernel import Module, Snippet
 
 
 class OutOfResourcesError(Exception):