+++ test/test_transformations.py
@@ -3,7 +3,7 @@ Test standard transformations
 """
 import pytest
 
-from reikna.cluda import Module
+from reikna.cluda import Snippet
 from reikna.elementwise import Elementwise, specialize_elementwise
 import reikna.transformations as tr
 
@@ -20,9 +20,8 @@ def pytest_generate_tests(metafunc):
         metafunc.parametrize('any_dtype', dtypes, ids=[str(x) for x in dtypes])
 
 
-identity = lambda output_dtype, input_dtype: Module.create(
-    lambda output, input: "${output.store}(idx, ${input.load}(idx));",
-    snippet=True)
+identity = lambda output_dtype, input_dtype: Snippet.create(
+    lambda output, input: "${output.store}(idx, ${input.load}(idx));")
 
 TestComputation = specialize_elementwise('output', 'input', None, identity)